# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ISC2 CPE Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Member {
  id: number;
  firstName: string;
  lastName: string;
  isc2Number: string;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  isc2Number: string;
}

export interface MemberUpdate {
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  /** @minLength 1 */
  isc2Number?: string;
}

export type EventGroupType = typeof EventGroupType[keyof typeof EventGroupType];


export const EventGroupType = {
  Group_A: 'Group A',
  Group_B: 'Group B',
} as const;

export interface Event {
  id: number;
  name: string;
  /** Event date in YYYY-MM-DD format */
  date: string;
  groupType: EventGroupType;
  cpeCredits: number;
  description: string;
  createdAt: string;
  attendeeCount: number;
}

export type EventInputGroupType = typeof EventInputGroupType[keyof typeof EventInputGroupType];


export const EventInputGroupType = {
  Group_A: 'Group A',
  Group_B: 'Group B',
} as const;

export interface EventInput {
  /** @minLength 1 */
  name: string;
  /** YYYY-MM-DD */
  date: string;
  groupType: EventInputGroupType;
  /**
     * @minimum 0.5
     * @maximum 10
     */
  cpeCredits: number;
  /** @maxLength 100 */
  description: string;
}

export type EventUpdateGroupType = typeof EventUpdateGroupType[keyof typeof EventUpdateGroupType];


export const EventUpdateGroupType = {
  Group_A: 'Group A',
  Group_B: 'Group B',
} as const;

export interface EventUpdate {
  /** @minLength 1 */
  name?: string;
  date?: string;
  groupType?: EventUpdateGroupType;
  /**
     * @minimum 0.5
     * @maximum 10
     */
  cpeCredits?: number;
  /** @maxLength 100 */
  description?: string;
}

export interface Attendee {
  memberId: number;
  firstName: string;
  lastName: string;
  isc2Number: string;
  checkedInAt: string;
}

export interface AttendeeInput {
  /** ID of existing member to check in */
  memberId: number;
}

export interface DashboardStats {
  totalMembers: number;
  totalEvents: number;
  groupACount: number;
  groupBCount: number;
  recentEvents: Event[];
}

export type ListMembersParams = {
/**
 * Search by name or ISC2 number
 */
search?: string;
};