# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ISC2 CPE Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: members
    description: ISC2 member directory
  - name: events
    description: CPE events
  - name: attendees
    description: Event attendance
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name or ISC2 number
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "409":
          description: ISC2 number already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: ISC2 number already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events with attendee counts
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}/attendees:
    get:
      operationId: listEventAttendees
      tags: [attendees]
      summary: List attendees for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of attendees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendee"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: checkInAttendee
      tags: [attendees]
      summary: Check in an attendee (existing or new member)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendeeInput"
      responses:
        "201":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}/attendees/{memberId}:
    delete:
      operationId: removeAttendee
      tags: [attendees]
      summary: Remove an attendee from an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}/export:
    get:
      operationId: exportEventExcel
      tags: [events]
      summary: Export event attendees as ISC2-formatted Excel file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Member:
      type: object
      required: [id, firstName, lastName, isc2Number, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        isc2Number:
          type: string
        createdAt:
          type: string

    MemberInput:
      type: object
      required: [firstName, lastName, isc2Number]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        isc2Number:
          type: string
          minLength: 1

    MemberUpdate:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        isc2Number:
          type: string
          minLength: 1

    Event:
      type: object
      required: [id, name, date, groupType, cpeCredits, description, createdAt, attendeeCount]
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
          description: Event date in YYYY-MM-DD format
        groupType:
          type: string
          enum: ["Group A", "Group B"]
        cpeCredits:
          type: number
        description:
          type: string
        createdAt:
          type: string
        attendeeCount:
          type: integer

    EventInput:
      type: object
      required: [name, date, groupType, cpeCredits, description]
      properties:
        name:
          type: string
          minLength: 1
        date:
          type: string
          description: YYYY-MM-DD
        groupType:
          type: string
          enum: ["Group A", "Group B"]
        cpeCredits:
          type: number
          minimum: 0.5
          maximum: 10
        description:
          type: string
          maxLength: 100

    EventUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        date:
          type: string
        groupType:
          type: string
          enum: ["Group A", "Group B"]
        cpeCredits:
          type: number
          minimum: 0.5
          maximum: 10
        description:
          type: string
          maxLength: 100

    Attendee:
      type: object
      required: [memberId, firstName, lastName, isc2Number, checkedInAt]
      properties:
        memberId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        isc2Number:
          type: string
        checkedInAt:
          type: string

    AttendeeInput:
      type: object
      required: [memberId]
      properties:
        memberId:
          type: integer
          description: ID of existing member to check in

    DashboardStats:
      type: object
      required: [totalMembers, totalEvents, groupACount, groupBCount, recentEvents]
      properties:
        totalMembers:
          type: integer
        totalEvents:
          type: integer
        groupACount:
          type: integer
        groupBCount:
          type: integer
        recentEvents:
          type: array
          items:
            $ref: "#/components/schemas/Event"