# lib/api-zod/src/generated/types/attendee.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ISC2 CPE Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Attendee {
  memberId: number;
  firstName: string;
  lastName: string;
  isc2Number: string;
  checkedInAt: string;
}