# lib/api-zod/src/generated/types/attendeeInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ISC2 CPE Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendeeInput {
  /** ID of existing member to check in */
  memberId: number;
}