# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ISC2 CPE Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from './event';

export interface DashboardStats {
  totalMembers: number;
  totalEvents: number;
  groupACount: number;
  groupBCount: number;
  recentEvents: Event[];
}