# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ISC2 CPE Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { EventGroupType } from './eventGroupType';

export interface Event {
  id: number;
  name: string;
  /** Event date in YYYY-MM-DD format */
  date: string;
  groupType: EventGroupType;
  cpeCredits: number;
  description: string;
  createdAt: string;
  attendeeCount: number;
}